refactor: change modal from hook to component
