+++ src/pages/donations/CausesPage/ChooseCauseModal/CauseImage/styles.ts
